refactor(api-file-manager): move shared upload utils to base package
Move FileKey, FileExtension, FileNormalizer, FileUploadModifier,
createFileNormalizerFromContext, checkPermissions, and mimeTypes from both
provider packages into packages/api-file-manager/src/features/upload/utils/.
Unify FileUploadModifierPlugin.type to "fm.uploadModifier". Move FileKey test
to packages/api-file-manager/__tests__/features/upload/utils/.
