+++ packages/api-file-manager-s3/package.json
@@ -28,11 +28,8 @@
     "@webiny/utils": "0.0.0",
     "@webiny/validation": "0.0.0",
     "exifreader": "^4.41.0",
-    "mime": "^4.1.0",
     "object-hash": "^3.0.0",
     "p-map": "^7.0.4",
-    "p-reduce": "^3.0.0",
-    "sanitize-filename": "^1.6.4",
     "sharp": "^0.35.1"
   },
   "devDependencies": {
