+++ packages/api-file-manager-server/src/index.ts
@@ -11,7 +11,7 @@ import { FlushCacheFeature } from "~/features/FlushCache/feature.js";
 import { GetFileContentsByIdFeature } from "~/features/GetFileContentsById/feature.js";
 import { GetFileContentsByKeyFeature } from "~/features/GetFileContentsByKey/feature.js";
 import { WriteFileMetadataFeature } from "~/features/WriteFileMetadata/feature.js";
-export { createFileUploadModifier } from "./utils/FileUploadModifier.js";
+export { createFileUploadModifier } from "@webiny/api-file-manager/features/upload/index.js";
 export { createAssetDelivery } from "./assetDelivery/createAssetDelivery.js";
 
 const contextPlugin = new ContextPlugin(context => {
