--- packages/api-file-manager-server/src/utils/createFileNormalizerFromContext.ts
@@ -1,11 +1,0 @@
-import type { Context } from "@webiny/api/types.js";
-import { FileNormalizer } from "~/utils/FileNormalizer.js";
-import { createModifierFromPlugins, FileUploadModifierPlugin } from "~/utils/FileUploadModifier.js";
-
-export const createFileNormalizerFromContext = (context: Context) => {
-    const modifierPlugins = context.plugins.byType<FileUploadModifierPlugin>(
-        FileUploadModifierPlugin.type
-    );
-
-    return new FileNormalizer(createModifierFromPlugins(modifierPlugins));
-};
