+++ packages/api-file-manager/package.json
@@ -32,8 +32,12 @@
     "@webiny/handler": "0.0.0",
     "@webiny/handler-graphql": "0.0.0",
     "@webiny/plugins": "0.0.0",
+    "@webiny/utils": "0.0.0",
     "@webiny/wcp": "0.0.0",
     "cache-control-parser": "^2.2.0",
+    "mime": "^4.1.0",
+    "p-reduce": "^3.0.0",
+    "sanitize-filename": "^1.6.4",
     "zod": "4.4.3"
   },
   "devDependencies": {
