+++ packages/api-file-manager/src/features/upload/index.ts
@@ -3,3 +3,13 @@ export type { FileData } from "./types.js";
 export type { UploadPayloadResponse } from "./types.js";
 export type { MultiPartUploadFilePart } from "./types.js";
 export type { CreateMultiPartUploadResult } from "./types.js";
+
+export { checkPermissions } from "./utils/checkPermissions.js";
+export { FileKey } from "./utils/FileKey.js";
+export { FileExtension } from "./utils/FileExtension.js";
+export { mimeTypes } from "./utils/mimeTypes.js";
+export { FileNormalizer } from "./utils/FileNormalizer.js";
+export { createFileUploadModifier } from "./utils/FileUploadModifier.js";
+export { FileUploadModifierPlugin } from "./utils/FileUploadModifier.js";
+export { createModifierFromPlugins } from "./utils/FileUploadModifier.js";
+export { createFileNormalizerFromContext } from "./utils/createFileNormalizerFromContext.js";
