+++ packages/api-file-manager/src/features/upload/utils/createFileNormalizerFromContext.ts
@@ -1,6 +1,7 @@
 import type { Context } from "@webiny/api/types.js";
-import { FileNormalizer } from "~/utils/FileNormalizer.js";
-import { createModifierFromPlugins, FileUploadModifierPlugin } from "~/utils/FileUploadModifier.js";
+import { FileNormalizer } from "./FileNormalizer.js";
+import { createModifierFromPlugins } from "./FileUploadModifier.js";
+import { FileUploadModifierPlugin } from "./FileUploadModifier.js";
 
 export const createFileNormalizerFromContext = (context: Context) => {
     const modifierPlugins = context.plugins.byType<FileUploadModifierPlugin>(
