+++ packages/api-file-manager/src/features/upload/utils/mimeTypes.ts
@@ -1,7 +1,7 @@
 import vendorTypes from "mime/types/other.js";
 import standardTypes from "mime/types/standard.js";
 
-/**
+/*
  * We need to get all extensions that are defined for a specific content type.
  * This operation is not available via the `mime` package, so we create such an object ourselves.
  */
