+++ packages/api-file-manager/tsconfig.json
@@ -11,9 +11,9 @@
     { "path": "../handler" },
     { "path": "../handler-graphql" },
     { "path": "../plugins" },
+    { "path": "../utils" },
     { "path": "../wcp" },
-    { "path": "../handler-aws" },
-    { "path": "../utils" }
+    { "path": "../handler-aws" }
   ],
   "compilerOptions": {
     "rootDirs": ["./src", "./__tests__"],
@@ -44,12 +44,12 @@
       "@webiny/handler-graphql": ["../handler-graphql/src"],
       "@webiny/plugins/*": ["../plugins/src/*"],
       "@webiny/plugins": ["../plugins/src"],
+      "@webiny/utils/*": ["../utils/src/*"],
+      "@webiny/utils": ["../utils/src"],
       "@webiny/wcp/*": ["../wcp/src/*"],
       "@webiny/wcp": ["../wcp/src"],
       "@webiny/handler-aws/*": ["../handler-aws/src/*"],
-      "@webiny/handler-aws": ["../handler-aws/src"],
-      "@webiny/utils/*": ["../utils/src/*"],
-      "@webiny/utils": ["../utils/src"]
+      "@webiny/handler-aws": ["../handler-aws/src"]
     }
   }
 }
